refactor and resolve code smell
