+++ packages/web-service/src/pages/applications/application-summary.js
@@ -1,45 +1,13 @@
 import pageRoute from '../../routes/page-route.js'
-import { APPLICATION_SUMMARY, APPLICATIONS } from '../../uris.js'
-import { DEFAULT_ROLE } from '../../constants.js'
+import { APPLICATION_SUMMARY } from '../../uris.js'
 import { APIRequests } from '../../services/api-requests.js'
-import { PowerPlatformKeys } from '@defra/wls-powerapps-keys'
 import { timestampFormatter } from '../common/common.js'
 import { ContactRoles } from '../contact/common/contact-roles.js'
 import { addressLine } from '../service/address.js'
-
-const { BACKEND_STATUS, APPLICATION_TYPES } = PowerPlatformKeys
-
-export const checkData = async (request, h) => {
-  const params = new URLSearchParams(request.query)
-  if (!params.get('applicationId')) {
-    return h.redirect(APPLICATIONS.uri)
-  }
-
-  const applicationId = params.get('applicationId')
-  const { userId } = await request.cache().getData()
-  const roles = await APIRequests.APPLICATION.findRoles(userId, applicationId)
-  if (!roles?.includes(DEFAULT_ROLE)) {
-    return h.redirect(APPLICATIONS.uri)
-  }
-
-  return null
-}
-
-// values to keys and keys to values
-export const statuses = Object.entries(BACKEND_STATUS)
-  .map(([k, v]) => ({ [v]: k }))
-  .reduce((p, c) => ({ ...p, ...c }))
-
-const findApplicationType = application => Object.entries(APPLICATION_TYPES)
-  .find(([_k, v]) => v === application.applicationTypeId)[0]
+import { statuses, checkData, getApplicationData } from './application-common-functions.js'
 
 export const getData = async request => {
-  const params = new URLSearchParams(request.query)
-  const applicationId = params.get('applicationId')
-  const journeyData = await request.cache().getData()
-  await request.cache().setData(Object.assign(journeyData, { applicationId }))
-  const application = await APIRequests.APPLICATION.getById(applicationId)
-  const applicationType = findApplicationType(application)
+  const { application, applicationType, applicationId } = await getApplicationData(request)
   const sites = await APIRequests.SITE.findByApplicationId(applicationId)
   const siteAddress = sites.length ? addressLine(sites[0]) : ''
   Object.assign(application, { applicationType, siteAddress })
